GNG-570: Move Contact Information fields to separate file
